# project/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Auditable Agent Runtime API
  version: 1.0.0
  description: Minimum MVP contract for a multi-tenant auditable Agent Runtime.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { description: Tokens refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
...
  responses:
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    ValidationError: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    RefreshRequest:
      type: object
      required: [refreshToken]
      additionalProperties: false
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 256, writeOnly: true }
    AuthResponse:
      type: object
      required: [data, requestId]
      properties:
        data: { type: object, required: [accessToken, expiresAt], properties: { accessToken: { type: string, writeOnly: true }, expiresAt: { type: string, format: date-time } } }
        requestId: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, VERSION_CONFLICT, RATE_LIMIT_EXCEEDED, QUOTA_EXCEEDED, BUDGET_EXCEEDED, INVALID_STATE_TRANSITION, APPROVAL_ALREADY_HANDLED, RESUME_WINDOW_EXPIRED, MODEL_PROVIDER_ERROR, TOOL_EXECUTION_FAILED, SANDBOX_TIMEOUT, INTERNAL_ERROR] }
            message: { type: string, maxLength: 500 }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...